+++ src/python/pants/engine/legacy/BUILD
@@ -28,6 +28,7 @@ python_library(
     'src/python/pants/engine:mapper',
     'src/python/pants/engine:objects',
     'src/python/pants/engine:parser',
+    'src/python/pants/engine:build_files',
     'src/python/pants/util:memo',
   ],
   tags = {"partially_type_checked"},
