+++ tests/python/pants_test/engine/examples/parsers.py
@@ -62,7 +62,7 @@ class JsonParser(Parser):
         )
         return JSONDecoder(object_hook=decoder, strict=True)
 
-    def parse(self, filepath, filecontent):
+    def parse(self, filepath, filecontent, load_statements=None):
         """Parse the given json encoded string into a list of top-level objects found.
 
         The parser accepts both blank lines and comment lines (those beginning with optional whitespace
@@ -71,6 +71,8 @@ class JsonParser(Parser):
         The parse also supports a simple protocol for serialized types that have an `_asdict` method.
         This includes `namedtuple` subtypes as well as any custom class with an `_asdict` method defined;
         see :class:`pants.engine.serializable.Serializable`.
+
+        Note: We ignore load statements in JSON Build files
         """
         json = ensure_text(filecontent)
 
@@ -225,7 +227,9 @@ class PythonAssignmentsParser(Parser):
             parse_globals[alias] = functools.partial(aliased, alias, symbol)
         return parse_globals
 
-    def parse(self, filepath, filecontent):
+    def parse(self, filepath, filecontent, load_statements=None):
+        # Note: This parser is used for testing, we ignore load statements.
+
         parse_globals = self._globals
 
         python = filecontent
@@ -287,7 +291,9 @@ class PythonCallbacksParser(Parser):
             parse_globals[alias] = functools.partial(registered, alias, symbol)
         return objects, parse_globals
 
-    def parse(self, filepath, filecontent):
+    def parse(self, filepath, filecontent, load_statements=None):
+        # Note: This parser is used for testing, we ignore load statements.
+
         objects, parse_globals = self._globals
 
         python = filecontent
